requirements.txt
dnspython
google-api-python-client
google-auth-oauthlib
lxml
psutil
pymongo
python-dotenv
python-magic
python-telegram-bot
requests
telegraph
tenacity
